[fault_resumable] Fix a nasty bug in fault_resumable_call [asm]
The problem was caused by the `custom_flags` field in regs: the asm func
fault_resumable_call() was not updated accordingly.

+++ include/tilck/kernel/arch/i386/arch_utils.h
@@ -38,7 +38,7 @@ typedef struct {
 
 static ALWAYS_INLINE int regs_intnum(regs *r)
 {
-   return (int)r->int_num;
+   return r->int_num;
 }
 
 static ALWAYS_INLINE void set_return_register(regs *r, uptr value)
